server: move the environment sample loop out of server.go
Release note: None
