+++ pkg/server/BUILD.bazel
@@ -20,6 +20,7 @@ go_library(
         "config_windows.go",
         "doc.go",
         "drain.go",
+        "env_sampler.go",
         "grpc_server.go",
         "gzip_response_writer.go",
         "index_usage_stats.go",
